Fixed Devel Generate page access callback was in the wrong file; standardized it to be like the other access calls Made the Tags submodule and the Flag integration submodule default views have the "created time" field link to the individual status update Made the individual status pages show comments Issue #1037780 - 2nd comment posts but doesn't appear by IceCreamYou
However, discovered a new problem: if a View with status comment forms has an AJAX pager, comments won't refresh correctly when a comment is saved on anything other than the first page. This is a huge problem and I'm not sure it's solveable without completely overriding Views' entire AJAX system.